+++ module-misc/src/test/java/com/arassec/igor/module/misc/action/util/FilterByRegExpActionTest.java
@@ -63,6 +63,14 @@ class FilterByRegExpActionTest extends MiscActionBaseTest {
         processedData = action.process(data, new JobExecution());
 
         assertTrue(processedData.isEmpty());
+
+        // matches, but should be dropped:
+        action.setExpression("igor.*test");
+        action.setDropMatching(true);
+
+        processedData = action.process(data, new JobExecution());
+
+        assertTrue(processedData.isEmpty());
     }
 
 }